requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pandas==2.2.2
numpy==1.26.4
yfinance==0.2.43
backtesting==0.3.3
bokeh==3.4.1
python-telegram-bot==20.7
python-dotenv==1.0.1
